+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## [6.1.0] - 2026-01-??
 ### Added
 - Custom HTML elements <wms-music-score-view> and <wms-playback-buttons>.
+- Document can be undefined in RenderContext/MusicScoreView.
 
 ## [6.0.1] - 2025-12-31
 ### Fixed